feat(aws-messaging): allow disabling deleting messages for lambda

+++ packages/nestjs-aws-messaging/lib/aws-transport-strategy.ts
@@ -91,6 +91,10 @@ export class AwsTransportStrategy extends Server implements CustomTransportStrat
       return;
     }
 
+    if (!this.options.deleteHandled) {
+      return;
+    }
+
     try {
       await this.deleteMessageFromQueue(message);
     } catch (error) {
